requirements.txt
fastapi
uvicorn
python-dotenv
pydantic>=2
langchain-core>=0.3
langgraph>=0.2
ibm-watsonx-ai>=1.2.0
astrapy>=1.4.2
neo4j>=5.16.0
httpx>=0.27
tenacity>=8.2
requests
beautifulsoup4
pdfminer.six
python-docx
python-pptx
openpyxl
lxml
numpy>=1.21.0
scikit-learn>=1.0.0
pyyaml
pydantic-settings>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
structlog>=23.0.0
